# contracts/crm-pipeline/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CRM Pipeline Service",
    "description": "CRM Pipeline: leads, opportunities, pipeline stages, activities, and sales handoff.",
    "license": { "name": "" },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/crm-pipeline/activities": {
      "get": {
        "tags": ["Activities"],
        "summary": "List activities",
        "operationId": "list_activities",
        "parameters": [
          { "name": "lead_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "opportunity_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "List of activities", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Activity" } } } } }
        },
        "security": [{ "bearer": [] }]
      },
      "post": {
        "tags": ["Activities"],
        "summary": "Log an activity",
        "operationId": "log_activity",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateActivityRequest" } } } },
        "responses": {
          "201": { "description": "Activity logged", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Activity" } } } }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/activities/{id}": {
      "get": {
        "tags": ["Activities"],
        "summary": "Get an activity",
        "operationId": "get_activity",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Activity detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Activity" } } } }
        },
        "security": [{ "bearer": [] }]
      },
      "put": {
        "tags": ["Activities"],
        "summary": "Update an activity",
        "operationId": "update_activity",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateActivityRequest" } } } },
        "responses": { "200": { "description": "Activity updated" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/activities/{id}/complete": {
      "post": {
        "tags": ["Activities"],
        "summary": "Complete an activity",
        "operationId": "complete_activity",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Activity completed" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/activity-types": {
      "get": {
        "tags": ["Activities"],
        "summary": "List activity types",
        "operationId": "list_activity_types",
        "responses": {
          "200": { "description": "Activity types", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ActivityType" } } } } }
        },
        "security": [{ "bearer": [] }]
      },
      "post": {
        "tags": ["Activities"],
        "summary": "Create an activity type",
        "operationId": "create_activity_type",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateActivityTypeRequest" } } } },
        "responses": {
          "201": { "description": "Activity type created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ActivityType" } } } }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/activity-types/{code}": {
      "put": {
        "tags": ["Activities"],
        "summary": "Update an activity type",
        "operationId": "update_activity_type",
        "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateActivityTypeRequest" } } } },
        "responses": { "200": { "description": "Activity type updated" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/contacts/{party_contact_id}/attributes": {
      "get": {
        "tags": ["ContactRoles"],
        "summary": "Get contact role attributes",
        "operationId": "get_contact_attributes",
        "parameters": [{ "name": "party_contact_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Contact role attributes", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContactRoleAttributes" } } } }
        },
        "security": [{ "bearer": [] }]
      },
      "put": {
        "tags": ["ContactRoles"],
        "summary": "Set contact role attributes",
        "operationId": "set_contact_attributes",
        "parameters": [{ "name": "party_contact_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpsertContactRoleRequest" } } } },
        "responses": { "200": { "description": "Attributes upserted" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/status-labels": {
      "get": {
        "tags": ["Labels"],
        "summary": "List lead status labels",
        "operationId": "list_status_labels",
        "responses": { "200": { "description": "Labels", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Label" } } } } } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/source-labels": {
      "get": {
        "tags": ["Labels"],
        "summary": "List lead source labels",
        "operationId": "list_source_labels",
        "responses": { "200": { "description": "Labels", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Label" } } } } } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/type-labels": {
      "get": {
        "tags": ["Labels"],
        "summary": "List opportunity type labels",
        "operationId": "list_type_labels",
        "responses": { "200": { "description": "Labels", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Label" } } } } } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/priority-labels": {
      "get": {
        "tags": ["Labels"],
        "summary": "List opportunity priority labels",
        "operationId": "list_priority_labels",
        "responses": { "200": { "description": "Labels", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Label" } } } } } },
        "security": [{ "bearer": [] }]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Activity": {
        "type": "object",
        "required": ["id", "tenant_id", "activity_type_code", "subject", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string" },
          "activity_type_code": { "type": "string" },
          "subject": { "type": "string" },
          "lead_id": { "type": ["string", "null"], "format": "uuid" },
          "opportunity_id": { "type": ["string", "null"], "format": "uuid" },
          "party_id": { "type": ["string", "null"], "format": "uuid" },
          "party_contact_id": { "type": ["string", "null"], "format": "uuid" },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "due_at": { "type": ["string", "null"], "format": "date-time" },
          "completed_at": { "type": ["string", "null"], "format": "date-time" },
          "notes": { "type": ["string", "null"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateActivityRequest": {
        "type": "object",
        "required": ["activity_type_code", "subject"],
        "properties": {
          "activity_type_code": { "type": "string" },
          "subject": { "type": "string" },
          "lead_id": { "type": ["string", "null"], "format": "uuid" },
          "opportunity_id": { "type": ["string", "null"], "format": "uuid" },
          "party_id": { "type": ["string", "null"], "format": "uuid" },
          "party_contact_id": { "type": ["string", "null"], "format": "uuid" },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "due_at": { "type": ["string", "null"], "format": "date-time" },
          "notes": { "type": ["string", "null"] }
        }
      },
      "UpdateActivityRequest": {
        "type": "object",
        "properties": {
          "subject": { "type": ["string", "null"] },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "due_at": { "type": ["string", "null"], "format": "date-time" },
          "notes": { "type": ["string", "null"] }
        }
      },
      "ActivityType": {
        "type": "object",
        "required": ["id", "tenant_id", "type_code", "display_label", "is_active", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string" },
          "type_code": { "type": "string" },
          "display_label": { "type": "string" },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateActivityTypeRequest": {
        "type": "object",
        "required": ["type_code", "display_label"],
        "properties": {
          "type_code": { "type": "string" },
          "display_label": { "type": "string" }
        }
      },
      "UpdateActivityTypeRequest": {
        "type": "object",
        "properties": {
          "display_label": { "type": ["string", "null"] },
          "is_active": { "type": ["boolean", "null"] }
        }
      },
      "ContactRoleAttributes": {
        "type": "object",
        "required": ["id", "tenant_id", "party_contact_id", "is_decision_maker", "is_active", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string" },
          "party_contact_id": { "type": "string", "format": "uuid" },
          "sales_role": { "type": ["string", "null"] },
          "influence_level": { "type": ["string", "null"] },
          "is_decision_maker": { "type": "boolean" },
          "notes": { "type": ["string", "null"] },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "UpsertContactRoleRequest": {
        "type": "object",
        "properties": {
          "sales_role": { "type": ["string", "null"] },
          "influence_level": { "type": ["string", "null"] },
          "is_decision_maker": { "type": ["boolean", "null"] },
          "notes": { "type": ["string", "null"] }
        }
      },
      "Label": {
        "type": "object",
        "required": ["canonical_value", "display_label", "sort_order", "is_active"],
        "properties": {
          "canonical_value": { "type": "string" },
          "display_label": { "type": "string" },
          "sort_order": { "type": "integer" },
          "is_active": { "type": "boolean" }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}